@@ -2,11 +2,12 @@
 
 """
 
-from mindsdb import *
+from mindsdb import Predictor
 
 # Here we use the model to make predictions (NOTE: You need to run train.py first)
-result = MindsDB().predict(predict='Main_Engine_Fuel_Consumption_MT_day', model_name='fuel', from_data = 'fuel_predict.csv')
+result = Predictor(name='fuel').predict(when_data = 'fuel_predict.csv')
 
 # you can now print the results
 print('The predicted main engine fuel consumption')
-print(result.predicted_values)
+for row in result:
+  print(row)
