@@ -1,6 +1,8 @@
 # Cochran’s sample size calculator
+from scipy.stats import norm
 
-def sampleSize(
+
+def calculate_sample_size(
     population_size,
     margin_error=.05,
     confidence_level=.99,
@@ -48,7 +50,9 @@ def sampleSize(
     if confidence_level in zdict:
         z = zdict[confidence_level]
     else:
-        from scipy.stats import norm
+        #Inf fix
+        if alpha == 0.0:
+            alpha += 0.001
         z = norm.ppf(1 - (alpha/2))
     N = population_size
     M = margin_error
