@@ -1,50 +1,20 @@
-import pandas
-import logging
-import csv
-import mindsdb.config as CONFIG
 import re
-import urllib3
 from io import BytesIO, StringIO
 import csv
 import codecs
 import json
 import traceback
+import codecs
+
+import pandas as pd
+from pandas.io.json import json_normalize
+import requests
 
 from mindsdb.libs.data_types.data_source import DataSource
-from pandas.io.json import json_normalize
+from mindsdb.libs.data_types.mindsdb_logger import log
+
 
 class FileDS(DataSource):
-
-    def clean(self, header):
-
-        clean_header = []
-        col_count={}
-
-        replace_chars = """ ,./;'[]!@#$%^&*()+{-=+~`}\\|:"<>?"""
-
-        for col in header:
-            orig_col = col
-            for char in replace_chars:
-                col = col.replace(char,'_')
-            col = re.sub('_+','_',col)
-            if col[-1] == '_':
-                col = col[:-1]
-            col_count[col] = 1 if col not in col_count else col_count[col]+1
-            if col_count[col] > 1:
-                col = col+'_'+str(col_count[col])
-
-            if orig_col != col:
-                logging.debug('[Column renamed] {orig_col} to {col}'.format(orig_col=orig_col, col=col))
-
-            self._col_map[orig_col] = col
-            clean_header.append(col)
-
-        if clean_header != header:
-            string = """\n    {cols} \n""".format(cols=",\n    ".join(clean_header))
-            logging.debug('The Columns have changed, here are the renamed columns: \n {string}'.format(string=string))
-
-
-        return  clean_header
 
     def cleanRow(self, row):
         n_row = []
@@ -58,7 +28,6 @@ class FileDS(DataSource):
     def _getDataIo(self, file):
         """
         This gets a file either url or local file and defiens what the format is as well as dialect
-
         :param file: file path or url
         :return: data_io, format, dialect
         """
@@ -71,11 +40,10 @@ class FileDS(DataSource):
 
         # get data from either url or file load in memory
         if file[:5] == 'http:' or file[:6] == 'https:':
-            urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
-            http = urllib3.PoolManager()
-            r = http.request('GET', file, preload_content=False)
-
-            data.write(r.read())
+            r = requests.get(file, stream=True)
+            if r.status_code == 200:
+                for chunk in r:
+                    data.write(chunk)
             data.seek(0)
 
         # else read file from local file system
@@ -84,7 +52,7 @@ class FileDS(DataSource):
                 data = open(file, 'rb')
             except Exception as e:
                 error = 'Could not load file, possible exception : {exception}'.format(exception = e)
-                logging.error(error)
+                log.error(error)
                 raise ValueError(error)
 
 
@@ -99,7 +67,7 @@ class FileDS(DataSource):
         xlsx_sig2 = b'\x50\x4B\x03\x04'
         xls_sig = b'\x09\x08\x10\x00\x00\x06\x05\x00'
 
-        # differnt whence, offset, size for different types
+        # different whence, offset, size for different types
         excel_meta = [ ('xls', 0, 512, 8), ('xlsx', 2, -22, 4)]
 
         for filename, whence, offset, size in excel_meta:
@@ -123,10 +91,15 @@ class FileDS(DataSource):
         byte_str = data.read()
         # Move it to StringIO
         try:
-            data = StringIO(byte_str.decode('UTF-8'))
+            # Handle Microsoft's BOM "special" UTF-8 encoding
+            if byte_str.startswith(codecs.BOM_UTF8):
+                data = StringIO(byte_str.decode('utf-8-sig'))
+            else:
+                data = StringIO(byte_str.decode('utf-8'))
+
         except:
-            logging.error(traceback.format_exc())
-            logging.error('Could not load into string')
+            log.error(traceback.format_exc())
+            log.error('Could not load into string')
 
         # see if its JSON
         buffer = data.read(100)
@@ -148,10 +121,18 @@ class FileDS(DataSource):
         # lets try to figure out if its a csv
         try:
             data.seek(0)
-            full = len(data.read())
-            data.seek(0)
-            bytes_to_read = int(full*0.3)
-            dialect = csv.Sniffer().sniff(data.read(bytes_to_read))
+            first_few_lines = []
+            i = 0
+            for line in data:
+                if line in ['\r\n','\n']:
+                    continue
+                first_few_lines.append(line)
+                i += 1
+                if i > 0:
+                    break
+
+            accepted_delimiters = [',','\t', ';']
+            dialect = csv.Sniffer().sniff(''.join(first_few_lines[0]), delimiters=accepted_delimiters)
             data.seek(0)
             # if csv dialect identified then return csv
             if dialect:
@@ -160,44 +141,41 @@ class FileDS(DataSource):
                 return data, None, dialect
         except:
             data.seek(0)
-            logging.error('Could not detect format for this file')
-            logging.error(traceback.format_exc())
+            log.error('Could not detect format for this file')
+            log.error(traceback.format_exc())
             # No file type identified
             return data, None, dialect
 
 
 
 
-    def _setup(self,file, clean_header = True, clean_rows = True, custom_parser = None):
+    def _setup(self,file, clean_rows = True, custom_parser = None):
         """
         Setup from file
         :param file: fielpath or url
-        :param clean_header: if you want to clean header column names
         :param clean_rows:  if you want to clean rows for strange null values
         :param custom_parser: if you want to parse the file with some custom parser
-
         """
 
+        col_map = {}
         # get file data io, format and dialect
         data, format, dialect = self._getDataIo(file)
         data.seek(0) # make sure we are at 0 in file pointer
 
         if format is None:
-            logging.error('Could not laod file into any format, supported formats are csv, json, xls, xslx')
+            log.error('Could not laod file into any format, supported formats are csv, json, xls, xslx')
 
         if custom_parser:
             header, file_data = custom_parser(data, format)
 
         elif format == 'csv':
-
             csv_reader = list(csv.reader(data, dialect))
             header = csv_reader[0]
             file_data =  csv_reader[1:]
 
-
         elif format in ['xlsx', 'xls']:
             data.seek(0)
-            df = pandas.read_excel(data)
+            df = pd.read_excel(data)
             header = df.columns.values.tolist()
             file_data = df.values.tolist()
 
@@ -208,8 +186,8 @@ class FileDS(DataSource):
             header = df.columns.values.tolist()
             file_data = df.values.tolist()
 
-        if clean_header == True:
-            header = self.clean(header)
+        for col in header:
+            col_map[col] = col
 
         if clean_rows == True:
             file_list_data = []
@@ -219,9 +197,7 @@ class FileDS(DataSource):
         else:
             file_list_data = file_data
 
-        self.setDF(pandas.DataFrame(file_list_data, columns=header))
-
-
-
-
-
+        try:
+            return pd.DataFrame(file_list_data, columns=header), col_map
+        except:
+            return pd.read_csv(file), col_map
