@@ -27,10 +27,6 @@ import time
 
 
 class BaseModule():
-
-    phase_name = PHASE_END
-    log_on_run = True
-
     def __init__(self, session, transaction, **kwargs):
         '''
         Initialize the base module and the basic global variables
@@ -44,10 +40,7 @@ class BaseModule():
         self.transaction = transaction # type: controllers.transaction_controller.TransactionController
         self.output = {}
         self.setup(**kwargs)
-        self.logging = self.session.logging
-
-    def loadPhaseModule(self, module_name):
-        return self.transaction.getPhaseInstance(module_name)
+        self.log = self.transaction.log
 
     def run(self):
         pass
@@ -56,24 +49,12 @@ class BaseModule():
         start = time.time()
         class_name = type(self).__name__
 
-        # log warning if no phase name has been set or designed for the model
-        if self.phase_name == PHASE_END and self.log_on_run:
-            self.session.logging.error('Module {class_name} has no \'phase_name\' defined and therefore it cannot be properly tested'.format(class_name=class_name))
+        self.log.info('[START] {class_name}'.format(class_name=class_name))
 
-        if self.log_on_run:
-            self.session.logging.info('[START] {class_name}'.format(class_name=class_name))
-
-        # if we are past the breakpoint do nothing, breakpoints are used when testing a particular module
-        if self.transaction.breakpoint is not None and self.phase_name > self.transaction.breakpoint:
-            self.session.logging.warning('Module {class_name} has a phase that is beyond breakpoint, module\'s phase: {current_phase}, transaction\'s breakpoint: {breakpoint}'.format(class_name=class_name, current_phase=self.phase_name, breakpoint=self.transaction.breakpoint))
-            #exit()
-            return self.output
-
-        # else run it
         ret = self.run(**kwargs)
         execution_time = time.time() - start
-        if self.log_on_run:
-            self.session.logging.info('[END] {class_name}, execution time: {execution_time:.3f} seconds'.format(class_name=class_name, execution_time=execution_time))
+
+        self.log.info('[END] {class_name}, execution time: {execution_time:.3f} seconds'.format(class_name=class_name, execution_time=execution_time))
         return ret
 
     def setup(self, **kwargs):
